# services/api-gateway/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: apipratudo API Gateway
  version: 0.2.0
  description: API Gateway como front door para rotas /v1 do apipratudo.
servers:
  - url: https://api.apipratudo.local
security:
  - ApiKeyAuth: []
...
paths:
...
  /v1/servicos:
    get:
      summary: Listar servicos cadastrados
      tags:
        - agendamento
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Lista de servicos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceResponse"
        "400":
          description: Erro de validacao
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: API key invalida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Cota esgotada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Criar servico de atendimento
      tags:
        - agendamento
      security:
        - ApiKeyAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Chave para idempotencia. Mesmo payload retorna a mesma resposta 200.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceCreateRequest"
      responses:
        "200":
          description: Servico criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponse"
        "400":
          description: Erro de validacao
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: API key invalida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Cota esgotada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflito de idempotencia
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/agenda:
    get:
      summary: Consultar agenda por periodo
      tags:
        - agendamento
      security:
        - ApiKeyAuth: []
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: agendaId
          in: query
          required: false
          schema:
            type: string
            default: main
      responses:
        "200":
          description: Agenda retornada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgendaResponse"
        "400":
          description: Erro de validacao
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: API key invalida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Cota esgotada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/slots-disponiveis:
    get:
      summary: Listar slots disponiveis por dia
      tags:
        - agendamento
      security:
        - ApiKeyAuth: []
      parameters:
        - name: serviceId
          in: query
          required: true
          schema:
            type: string
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: agendaId
          in: query
          required: false
          schema:
            type: string
            default: main
        - name: zoneId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Slots disponiveis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlotsResponse"
        "400":
          description: Erro de validacao
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: API key invalida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Cota esgotada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/reservar:
    post:
      summary: Reservar horario
      tags:
        - agendamento
      security:
        - ApiKeyAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Chave para idempotencia. Mesmo payload retorna a mesma resposta 200.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationRequest"
      responses:
        "200":
          description: Reserva criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationResponse"
        "400":
          description: Erro de validacao
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: API key invalida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Cota esgotada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Slot indisponivel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/confirmar:
    post:
      summary: Confirmar reserva
      tags:
        - agendamento
      security:
        - ApiKeyAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Chave para idempotencia. Mesmo payload retorna a mesma resposta 200.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmRequest"
      responses:
        "200":
          description: Reserva confirmada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmResponse"
        "400":
          description: Erro de validacao
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: API key invalida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Cota esgotada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Reserva expirada ou indisponivel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/cancelar:
    post:
      summary: Cancelar reserva
      tags:
        - agendamento
      security:
        - ApiKeyAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Chave para idempotencia. Mesmo payload retorna a mesma resposta 200.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelRequest"
      responses:
        "200":
          description: Reserva cancelada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelResponse"
        "400":
          description: Erro de validacao
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: API key invalida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Cota esgotada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflito de idempotencia
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/notificar:
    post:
      summary: Registrar notificacao
      tags:
        - agendamento
      security:
        - ApiKeyAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Chave para idempotencia. Mesmo payload retorna a mesma resposta 200.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotifyRequest"
      responses:
        "200":
          description: Notificacao registrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotifyResponse"
        "400":
          description: Erro de validacao
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: API key invalida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Cota esgotada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflito de idempotencia
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - details
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Validation error
        details:
          type: array
          items:
            type: string
          example:
            - targetUrl must not be blank
        traceId:
          type: string
          example: 4bf92f3577b34da6a3ce929d0e0e4736
...
    ServiceCreateRequest:
      type: object
      required:
        - name
        - durationMin
        - prepMin
        - bufferMin
        - noShowFeeCents
        - active
      properties:
        name:
          type: string
          example: Corte Masculino
        durationMin:
          type: integer
          example: 40
        prepMin:
          type: integer
          example: 5
        bufferMin:
          type: integer
          example: 10
        noShowFeeCents:
          type: integer
          format: int64
          example: 2000
        active:
          type: boolean
          example: true
    ServiceResponse:
      type: object
      required:
        - id
        - name
        - durationMin
        - prepMin
        - bufferMin
        - noShowFeeCents
        - active
        - createdAt
      properties:
        id:
          type: string
          example: srv_123
        name:
          type: string
        durationMin:
          type: integer
        prepMin:
          type: integer
        bufferMin:
          type: integer
        noShowFeeCents:
          type: integer
          format: int64
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CustomerRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: Ana
        phone:
          type: string
          example: "+5588999990000"
        email:
          type: string
          example: ana@email.com
    ReservationRequest:
      type: object
      required:
        - serviceId
        - agendaId
        - startAt
        - customer
      properties:
        serviceId:
          type: string
          example: srv_123
        agendaId:
          type: string
          example: main
        startAt:
          type: string
          format: date-time
          example: "2026-01-26T12:00:00Z"
        customer:
          $ref: "#/components/schemas/CustomerRequest"
        notes:
          type: string
          example: Preferencia por tesoura.
    ReservationResponse:
      type: object
      required:
        - appointmentId
        - status
        - holdExpiresAt
        - startAt
        - endAt
      properties:
        appointmentId:
          type: string
          example: apt_123
        status:
          type: string
          example: HELD
        holdExpiresAt:
          type: string
          format: date-time
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
    ConfirmRequest:
      type: object
      required:
        - appointmentId
      properties:
        appointmentId:
          type: string
          example: apt_123
    ConfirmResponse:
      type: object
      required:
        - appointmentId
        - status
      properties:
        appointmentId:
          type: string
        status:
          type: string
          example: CONFIRMED
    CancelRequest:
      type: object
      required:
        - appointmentId
      properties:
        appointmentId:
          type: string
          example: apt_123
        reason:
          type: string
          example: Cliente pediu remarcacao
    CancelResponse:
      type: object
      required:
        - appointmentId
        - status
        - cancellationFeeCents
      properties:
        appointmentId:
          type: string
        status:
          type: string
          example: CANCELLED
        cancellationFeeCents:
          type: integer
          format: int64
          example: 0
    NotifyRequest:
      type: object
      required:
        - appointmentId
        - type
      properties:
        appointmentId:
          type: string
          example: apt_123
        type:
          type: string
          example: REMINDER
    NotifyResponse:
      type: object
      required:
        - ok
        - appointmentId
        - type
      properties:
        ok:
          type: boolean
          example: true
        appointmentId:
          type: string
        type:
          type: string
    SlotResponse:
      type: object
      required:
        - startAt
        - endAt
      properties:
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
    SlotsResponse:
      type: object
      required:
        - serviceId
        - date
        - agendaId
        - slotStepMin
        - slots
      properties:
        serviceId:
          type: string
        date:
          type: string
          format: date
        agendaId:
          type: string
        slotStepMin:
          type: integer
        slots:
          type: array
          items:
            $ref: "#/components/schemas/SlotResponse"
    AppointmentResponse:
      type: object
      required:
        - appointmentId
        - serviceId
        - agendaId
        - status
        - startAt
        - endAt
      properties:
        appointmentId:
          type: string
        serviceId:
          type: string
        agendaId:
          type: string
        status:
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        holdExpiresAt:
          type: string
          format: date-time
        customer:
          $ref: "#/components/schemas/CustomerRequest"
        notes:
          type: string
    AgendaResponse:
      type: object
      required:
        - agendaId
        - from
        - to
        - appointments
      properties:
        agendaId:
          type: string
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/AppointmentResponse"